+++ ninja-core/src/main/java/ninja/Context.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2012-2016 the original author or authors.
+ * Copyright (C) 2012-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -425,14 +425,14 @@ public interface Context {
     /**
      * Get the (first) request header with the given name
      * 
-     * @return The header value
+     * @return The header value or null if not found.
      */
     String getHeader(String name);
 
     /**
      * Get all the request headers with the given name.
      *
-     * @return the header values
+     * @return the header values. Returns empty list if none are found.
      */
     List<String> getHeaders(String name);