+++ ninja-core/src/main/java/ninja/validation/Required.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2012-2016 the original author or authors.
+ * Copyright (C) 2012-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,11 +24,13 @@ import java.lang.annotation.Target;
 /**
  * Validates that the length of the field meets the given length constraints
  *
+ * @deprecated Please use Optional in controller method instead. 
  * @author James Roper
  */
 @WithValidator(Validators.RequiredValidator.class)
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.PARAMETER)
+@Deprecated
 public @interface Required {
     /**
      * The key for the violation message