+++ ninja-core/src/test/java/ninja/ReverseRouterTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2012-2016 the original author or authors.
+ * Copyright (C) 2012-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,7 +29,9 @@ import ninja.params.Param;
 import ninja.params.ParamParsers;
 import ninja.utils.MethodReference;
 import static org.hamcrest.CoreMatchers.is;
+import org.mockito.Mockito;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class ReverseRouterTest {
 
@@ -48,7 +50,9 @@ public class ReverseRouterTest {
         when(testControllerProvider.get()).thenReturn(new TestController());
         when(injector.getProvider(TestController.class)).thenReturn(testControllerProvider);
         when(injector.getInstance(ParamParsers.class)).thenReturn(new ParamParsers(Collections.emptySet()));
-        router = new RouterImpl(injector, ninjaProperties);
+        Provider<RouteBuilderImpl> routeBuilderImplProvider = mock(Provider.class);
+        when(routeBuilderImplProvider.get()).thenAnswer((invocation) -> new RouteBuilderImpl(ninjaProperties));
+        router = new RouterImpl(injector, ninjaProperties, routeBuilderImplProvider);
         reverseRouter = new ReverseRouter(ninjaProperties, router);
         
         router.GET().route("/home").with(TestController::home);