+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/src/main/java/dao/UserDao.java
@@ -3,7 +3,7 @@
 #set( $symbol_dollar = '$' )
 #set( $symbol_escape = '\' )
 /**
- * Copyright (C) 2012-2016 the original author or authors.
+ * Copyright (C) 2012-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.